bugfix for batched small-sieve start computations for logI>16

+++ sieve/las-smallsieve.cpp
@@ -389,8 +389,12 @@ void small_sieve_init(small_sieve_data_t & ssd,
                      *
                      * It is such that (B+c, 0) is in L_p.
                      *
-                     * */
-                    unsigned long offset = ((p - 1) << LOG_BUCKET_REGION) % p;
+                     * Since LOG_BUCKET_REGION is a priori something as
+                     * large as 16 bits, and p might also be of roughly
+                     * the same size, maybe even more (up to logI), we
+                     * must pay attention to overflows.
+                     */
+                    unsigned long offset = ((int64_t) (p - 1) << LOG_BUCKET_REGION) % p;
                     ssd_offsets.right.push_back(offset);
                 }
             } else {
